Shifted model to user_models
