+++ core/storage/skill/gae_models.py
@@ -150,46 +150,3 @@ class SkillSummaryModel(base_models.BaseModel):
     # model was created).
     skill_model_created_on = ndb.DateTimeProperty(required=True, indexed=True)
     version = ndb.IntegerProperty(required=True)
-
-
-class UserSkillMasteryModel(base_models.BaseModel):
-    """Model for storing a user's degree of mastery of a skill in Oppia.
-
-    This model stores the degree of mastery of each skill for a given user.
-
-    The id for this model is of form '{{USER_ID}}:{{SKILL_ID}}'
-    """
-
-    # The user id of the user.
-    user_id = ndb.StringProperty(required=True, indexed=True)
-    # The skill id for which the degree of mastery is stored.
-    skill_id = ndb.StringProperty(required=True, indexed=True)
-    # The degree of mastery of the user in the skill.
-    degree_of_mastery = ndb.FloatProperty(required=True, indexed=True)
-
-    @classmethod
-    def get_model_id(cls, user_id, skill_id):
-        """Returns model id corresponding to user and skill.
-
-        Args:
-            user_id: str. The user id of the logged in user.
-            skill_id: str. The unique id of the skill.
-
-        Returns:
-            str. The model id corresponding to user and skill.
-        """
-        return '%s:%s' % (user_id, skill_id)
-
-    @classmethod
-    def get_all_skill_mastery_models_for_user(cls, user_id):
-        """Returns all skill mastery models of a particular user.
-
-        Args:
-            user_id: str. The user id corresponding to a user.
-
-        Returns:
-            list. All skill mastery models where user_id belongs to the
-                entered user are returned as a list where each element
-                is one skill mastery model.
-        """
-        return cls.query(cls.user_id == user_id).fetch()
